% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_models.R
\name{batchSUVR}
\alias{batchSUVR}
\title{Calculate weighted SUVRs for specified ROIs in a participant batch}
\usage{
batchSUVR(participants, dir = "", tac_format = "PMOD",
  tac_file_suffix = ".tac", vol_format = "Voistat",
  vol_file_suffix = "_TAC.voistat", ROI_def, SUVR_def, reference, PVC,
  outfile = NULL)
}
\arguments{
\item{participants}{A vector of participant IDs}

\item{tac_format}{See loadTACfile()}

\item{tac_file_suffix}{This is how the ID corresponds to the TAC files}

\item{vol_format}{See loadVolumes()}

\item{vol_file_suffix}{This is how the ID corresponds to the volume files}

\item{ROI_def}{Object that defines combined ROIs, see ROI_definitions.R}

\item{SUVR_def}{is a vector of the start times for window to be used in SUVR}

\item{outfile}{Specify a filename to save the data.}

\item{corrected}{For PVC, true where the data is stored as _C in same tac file}

\item{corrected}{See calcSUVR() to determine whether this applies.}
}
\value{
A table of SUVR values for the specified ROIs for all participants.
}
\description{
For a vector of participant IDs and correspondingly named tac and volume
files, this calculates SUVR using calcSUVR and stores in a single table.
}
\details{
See calcSUVR() for how SUVR is calculated.
}
\examples{
batchSUVR(participants, ROI_def=standardROIs(),
          SUVR_def=c(3000, 3300, 3600, 3900), outfile="batch1.csv")
}
