% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tac.R
\name{calcTAC}
\alias{calcTAC}
\title{Calculate weighted time-activity curves for specified regions of interest}
\usage{
calcTAC(tac, volumes, ROI_def, merge, PVC)
}
\arguments{
\item{tac}{The time-activity curve data from loading function}

\item{volumes}{The ROI volume data from loading function}

\item{ROI_def}{The definition of ROIs by combining smaller ROIs from TAC file}

\item{merge}{If TRUE, includes the original ROIs in the output data}

\item{PVC}{If TRUE, appends "_C" to ROI name header (as in PMOD TAC files)}
}
\value{
Time-activity curves for the specified ROIs
}
\description{
Calculate weighted time-activity curves for specified regions of interest
}
