% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logan.R
\name{DVR_all_reference_Logan}
\alias{DVR_all_reference_Logan}
\title{Non-invasive reference Logan method for all ROIs in tac data}
\usage{
DVR_all_reference_Logan(tac_data, reference, k2prime, t_star = 0,
  method = "trapz")
}
\arguments{
\item{tac_data}{The time-activity curve data from calcTAC()}

\item{reference}{The reference region, e.g. "cerebellum"}

\item{k2prime}{A fixed value for k2' must be specified (e.g. 0.2)}

\item{t_star}{If 0, t* will be calculated using find_t_star()}

\item{method}{Method of inntegration, "trapz" or "integrate"}
}
\value{
Data frame with calculate DVRs for all ROIs
}
\description{
This calculates the DVR using the non-invasive reference Logan method for
all TACs in a supplied tac file. It uses DVR_reference_Logan.
}
