% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logan.R
\name{DVR_reference_Logan}
\alias{DVR_reference_Logan}
\title{Non-invasive reference Logan method}
\usage{
DVR_reference_Logan(tac_data, target, ref, k2prime, t_star,
  method = "trapz")
}
\arguments{
\item{tac_data}{The time-activity curve data from calcTAC()}

\item{target}{The name of the target ROI, e.g. "frontal"}

\item{ref}{The reference region, e.g. "cerebellum"}

\item{k2prime}{A fixed value for k2' must be specified (e.g. 0.2)}

\item{t_star}{If 0, t* will be calculated using find_t_star()}

\item{method}{Method of inntegration, "trapz" or "integrate"}
}
\value{
Data frame with calculate DVRs for all ROIs
}
\description{
This calculates the coefficient from the non-invasive Logan method, which
is equal to DVR.
}
