% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROI_definitions.R
\name{fullROIs}
\alias{fullROIs}
\title{Return a list of larger ROIs made up of the ROIs in the Hammer's atlas.}
\usage{
fullROIs()
}
\value{
A list of lists, where each list is an ROI (e.g.) frontal lobe that 
specifies the atomic ROIs from the atlas that make it up.
}
\description{
This includes the cortical regions of standardROIs() but also other regions.
It can be modified to suit the user's needs.
}
\examples{
standardROIs()
}
