% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_voistat}
\alias{load_voistat}
\title{Loads model data from file for use by other functions.}
\usage{
load_voistat(voistat_file, ROI_def, model = "VALUE")
}
\arguments{
\item{voistat_file}{Filename (e.g. participant_logan.voistat)}

\item{ROI_def}{The definition of ROIs by combining smaller ROIs from TAC file}

\item{model}{A string to name the variable being extracted, e.g. "Logan_DVR"}
}
\value{
data.frame with loaded model data in specified combined weighted ROIs
}
\description{
Loads model data from file for use by other functions.
}
