% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{loadTACfile}
\alias{loadTACfile}
\title{Loads TAC from file for use by other functions (default is PMOD .tac format)}
\usage{
loadTACfile(filename, format = "PMOD", acqtimes = NULL)
}
\arguments{
\item{filename}{(e.g. participant.TAC)}

\item{format}{(default, and only option currently, is .tac as from PMOD.}

\item{acqtimes}{File name for a .acqtimes file (as in PMOD)}
}
\value{
data.frame with loaded TAC data
}
\description{
Loads TAC from file for use by other functions (default is PMOD .tac format)
}
