% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{loadVolumes}
\alias{loadVolumes}
\title{Loads ROI volumes from file for use by other functions}
\usage{
loadVolumes(filename, format = "Voistat")
}
\arguments{
\item{filename}{(e.g. participant.voistat)}

\item{format}{(default is the TAC .voistat format from PMOD)}
}
\value{
data.frame with loaded TAC data
}
\description{
Loads ROI volumes from file for use by other functions
}
