% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEAR.R
\docType{package}
\name{PEAR}
\alias{PEAR}
\alias{PEAR-package}
\title{PEAR: PET Analysis in R}
\description{
Documentation placeholder -- describe the package.
continue...
}
\section{Foo functions}{

The foo functions ...
}

