% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ref_Logan}
\alias{plot_ref_Logan}
\title{Non-invasive reference Logan plot}
\usage{
plot_ref_Logan(tac_data, target, ref, k2prime, t_star = 0, error = 0.1,
  method = "trapz")
}
\arguments{
\item{tac_data}{The time-activity curve data from tac_roi()}

\item{target}{The name of the receptor-rich region, e.g. "frontal"}

\item{ref}{The reference region, e.g. "cerebellum"}

\item{k2prime}{A fixed value for k2' must be specified (e.g. 0.2)}

\item{t_star}{If 0, t* will be calculated using find_t_star()}

\item{error}{For find_t_star()}

\item{method}{Method of integration, "trapz" or "integrate"}
}
\value{
No return
f <- system.file("extdata", "AD06.tac", package="tacmagic")
fv <- system.file("extdata", "AD06_TAC.voistat", package="tacmagic")
AD06_tac <- load_tac(f, format="PMOD")
AD06_volume <- load_vol(fv, format="voistat")
AD06 <- tac_roi(tac=AD06_tac, volumes=AD06_volume, ROI_def=roi_ham_pib(),  
                merge=FALSE, PVC=FALSE)  

plot_ref_Logan(AD06, target="frontal", ref="cerebellum", 
               k2prime=0.2, t_star=0)
}
\description{
This plots the non-invasive Logan plot.
}
\seealso{
Other Logan plot functions: \code{\link{DVR_all_ref_Logan}},
  \code{\link{DVR_ref_Logan}}
}
\concept{Logan plot functions}
