% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logan.R
\name{plot_reference_Logan}
\alias{plot_reference_Logan}
\title{Non-invasive reference Logan plot}
\usage{
plot_reference_Logan(tac_data, target, ref, k2prime, t_star = 0,
  method = "trapz")
}
\arguments{
\item{tac_data}{The time-activity curve data from calcTAC()}

\item{ref}{The reference region, e.g. "cerebellum"}

\item{k2prime}{A fixed value for k2' must be specified (e.g. 0.2)}

\item{t_star}{If 0, t* will be calculated using find_t_star()}

\item{method}{Method of inntegration, "trapz" or "integrate"}
}
\value{
No return
}
\description{
This plots the non-invasive Logan plot.
}
