% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tac.R
\name{plot_tac}
\alias{plot_tac}
\title{Plots time activity curves from 1 or 2 participants or groups.}
\usage{
plot_tac(TACtable1, TACtable2 = NULL, ROIs, ymax = 25,
  time = "minutes", title = "")
}
\arguments{
\item{TACtable1}{(e.g. from tac_roi() or load_tac())}

\item{TACtable2}{An optional, second TAC, to plot for comparison}

\item{ROIs}{A vector of ROIs to plot, names matching the TAC headers}

\item{ymax}{The maximum value on the y-axis}

\item{time}{"seconds" or "minutes" depending on desired x-axis, converts tac}

\item{title}{A title for the plot}
}
\value{
Creates a plot
}
\description{
Plots time activity curves from 1 or 2 participants or groups.
}
\examples{
# f_raw_tac and f_raw_vol are the filenames of PMOD-generated files
f_raw_tac <- system.file("extdata", "AD06.tac", package="tacmagic") 
f_raw_vol <- system.file("extdata", "AD06_TAC.voistat", package="tacmagic")

tac <- load_tac(f_raw_tac)
vol <- load_vol(f_raw_vol)
AD06_tac_nc <- tac_roi(tac, vol, roi_ham_full(), merge=FALSE, PVC=FALSE)
plot_tac(AD06_tac_nc, ROIs=c("frontal", "cerebellum"), title="Example Plot")
}
\seealso{
Other tac functions: \code{\link{save_tac}},
  \code{\link{tac_roi}}
}
\concept{tac functions}
