% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\name{references}
\alias{references}
\title{Print and return a list of references for this package.}
\usage{
references()
}
\value{
List of references.
}
\description{
Print and return a list of references for this package.
}
\examples{
references()
}
