% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROI_definitions.R
\name{roi_AAL_full}
\alias{roi_AAL_full}
\title{Return a list of merged ROIs made up of the atomic ROIs in the AAL atlas.
The groupings for major lobes are as described in the PMOD documentation.}
\usage{
roi_AAL_full()
}
\value{
A list of lists, where each list is an ROI (e.g.) frontal lobe that
specifies the atomic ROIs from the atlas that make it up.
}
\description{
Return a list of merged ROIs made up of the atomic ROIs in the AAL atlas.
The groupings for major lobes are as described in the PMOD documentation.
}
\examples{
roi_AAL_full()
}
\references{
PMOD Groupings of the AAL Atlas.

Automated Anatomical Labeling of Activations in SPM Using a Macroscopic Anatomical Parcellation of the MNI MRI Single-Subject Brain. N. Tzourio-Mazoyer, B. Landeau, D. Papathanassiou, F. Crivello, O. Étard, N. Delcroix, B. Mazoyer, and M. Joliot. NeuroImage 2002, 15 :273-289 <doi:10.1006/nimg.2001.0978>
}
\seealso{
Other ROI definitions: 
\code{\link{roi_ham_full}()},
\code{\link{roi_ham_pib}()},
\code{\link{roi_ham_stand}()}
}
\concept{ROI definitions}
