% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROI_definitions.R
\name{roi_ham_full}
\alias{roi_ham_full}
\title{Return a list of larger ROIs made up of the ROIs in the Hammer's atlas.}
\usage{
roi_ham_full()
}
\value{
A list of lists, where each list is an ROI (e.g.) frontal lobe that 
specifies the atomic ROIs from the atlas that make it up.
}
\description{
This includes the cortical regions of roi_ham_stand() but also other regions.
It can be modified to suit the user's needs.
}
\examples{
roi_ham_full()
}
