% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROI_definitions.R
\name{roi_ham_pib}
\alias{roi_ham_pib}
\title{Return a list of merged ROIs made up of atomic ROIs in the Hammer's atlas.}
\usage{
roi_ham_pib()
}
\value{
A list of lists, where each list is an ROI (e.g.) frontal lobe that 
specifies the atomic ROIs from the atlas that make it up.
}
\description{
This includes the ROIs from roi_ham_full and also the PIB cortical composite
ROI as defined in the PMOD documentation and as widely used in PIB studies.
See PMOD Neuro Tool (PNEURO) (Version 4.0) documentation.
}
\examples{
roi_ham_pib()
}
\references{
Hammers, Alexander, Richard Allom, Matthias J. Koepp, Samantha L. Free, 
Ralph Myers, Louis Lemieux, Tejal N. Mitchell, David J. Brooks, and John S. 
Duncan. 2003. Three-dimensional Maximum Probability Atlas of the Human 
Brain, with Particular Reference to the Temporal Lobe. Human Brain Mapping 
19 (4): 224-247. doi:10.1002/hbm.10123
}
\seealso{
Other ROI definitions: \code{\link{roi_ham_full}},
  \code{\link{roi_ham_stand}}
}
\concept{ROI definitions}
