% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving.R
\name{save_tac}
\alias{save_tac}
\title{Save a tac object as a .tac file}
\usage{
save_tac(tac, outfile)
}
\arguments{
\item{tac}{The time-activity curve data, e.g. from load_tac() or tac_roi()}

\item{outfile}{The output filename}
}
\value{
Does not return an object, only saves a file
}
\description{
Saves a tac object, created by load_tac(), tac_roi() or manually, and 
saves it as a PMOD-formatted tac file. Using the .tac extension in the 
file name is recommended.
}
\seealso{
Other tac functions: 
\code{\link{plot.tac}()},
\code{\link{split_pvc}()},
\code{\link{tac_roi}()}
}
\concept{tac functions}
