% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUVR.R
\name{suvr}
\alias{suvr}
\title{Calculate weighted SUVRs for specified regions of interest}
\usage{
suvr(tac, SUVR_def, ref)
}
\arguments{
\item{tac}{The time-activity curve data from tac_roi()}

\item{SUVR_def}{a vector of start times for window to be used in SUVR}

\item{ref}{is a string, e.g. "cerbellum", to specify reference region}
}
\value{
A data.frame of SUVR values for the specified ROIs
}
\description{
Calculate the standardized uptake value ratio (SUVR) for all ROIs in the
provided tac data, using the specified reference region.
}
