% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUVR.R
\name{suvr_auc}
\alias{suvr_auc}
\title{Calculate SUVRs for regions of interest with AUC from mid-frame times}
\usage{
suvr_auc(tac, SUVR_def, ref)
}
\arguments{
\item{tac}{The time-activity curve data from tac_roi()}

\item{SUVR_def}{a vector of start times for window to be used in SUVR}

\item{ref}{is a string, e.g. "cerbellum", to specify reference region}
}
\value{
A data.frame of SUVR values for the specified ROIs
}
\description{
Calculate the standardized uptake value ratio (SUVR) for all ROIs in the
provided tac data, using the specified reference region. This is an alternate
to suvr() which should provide very similar values.
}
