% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tacmagic.R
\docType{package}
\name{tacmagic}
\alias{tacmagic}
\title{tacmagic: PET Analysis in R}
\description{
The main features of tacmagic are to load PET time activity curve (tac) data 
from multiple formats, merge ROIs weighted for volume, calculate binding 
potential models including SUVR and DVR, basic plotting, and calculation of 
cut-off values. Please see the walkthrough vignette for a detailed overview.
}
