% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batches.R
\name{tm_batch}
\alias{tm_batch}
\title{Calculate one or more models for a batch of participants}
\usage{
tm_batch(participants, models = c("SUVR", "Logan", "eslope"), PVC = F,
  dir = "", tac_format = "PMOD", tac_file_suffix = ".tac",
  vol_file_suffix = "_TAC.voistat", vol_format = "Voistat",
  ROI_def = NULL, SUVR_def = NULL, ref = "cerebellum", merge = F,
  k2prime = NULL, t_star = 0, master = NULL, outfile = NULL)
}
\arguments{
\item{participants}{A vector of participant IDs}

\item{models}{A vector of names of the models to calculate}

\item{PVC}{For PVC, true where the data is stored as _C in same tac file}

\item{dir}{A directory and/or file name prefix for the tac/volume files}

\item{tac_format}{Format of tac files provided: See loadTACfile()}

\item{tac_file_suffix}{How participant IDs corresponds to the TAC files}

\item{vol_file_suffix}{How participant IDs correspond to volume files}

\item{vol_format}{The file format that includes volumes: See loadVolumes()}

\item{ROI_def}{Object that defines combined ROIs, see ROI_definitions.R}

\item{SUVR_def}{is a vector of the start times for window to be used in SUVR}

\item{ref}{The name of the reference region for DVR/SUVR calculation}

\item{merge}{Passes value to calcTAC; T to keep original atomic ROIs}

\item{k2prime}{Fixed k2' for DVR calculation}

\item{t_star}{Change from 0 to manually specify a t* for DVR calculation}

\item{master}{Optionally, a data.frame of same format as return, to add to}

\item{outfile}{Specify a filename to save the data}
}
\value{
A table of SUVR values for the specified ROIs for all participants
}
\description{
For a vector of participant IDs and correspondingly named tac and volume
files, this produces SUVR, DVR, and upslope in a tidy data.frame. Current
model options are "SUVR", "Logan" and "eslope".
}
\details{
For further details about how the models are calculated, see the indiviudal
functions that they rely on. "SUVR" uses suvr(), "Logan" uses
DVR_all_reference_Logan(), and "eslope" uses peaksSlope().
}
