% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batches.R
\name{voistat_batch}
\alias{voistat_batch}
\title{Obtain values from voistat files (using load_voistat() for a batch.}
\usage{
voistat_batch(participants, ROI_def, dir = "", filesuffix, varname,
  otherdata = NULL, outfile)
}
\arguments{
\item{participants}{A vector of participant IDs}

\item{ROI_def}{Object that defines combined ROIs, see ROI_definitions.R}

\item{dir}{Directory and/or filename prefix of the files}

\item{filesuffix}{Optional filename characters between ID and ".voistat"}

\item{varname}{The name of the variable being exctracted, e.g. "SRTM"}

\item{otherdata}{A data.frame of the same participants to add the new data to}

\item{outfile}{Specify a filename to save the data}
}
\value{
A table of values for the specified ROIs for all participants.
}
\description{
For a vector of participant IDs and correspondingly named .voistat files,
this extracts the value from the files for the specified ROIs.
}
\details{
See load_voistat() for specifics.
}
