% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{voistatScraper}
\alias{voistatScraper}
\title{Loads model data from file for use by other functions.}
\usage{
voistatScraper(voistat_file, ROI_def, model = "VALUE")
}
\arguments{
\item{filename}{(e.g. participant_logan.voistat)}

\item{format}{(default is the TAC .voistat format from PMOD).}
}
\value{
data.frame with loaded model data in specified combined weighted ROIs.
}
\description{
Loads model data from file for use by other functions.
}
