% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSUVR.R
\name{calcSUVR}
\alias{calcSUVR}
\title{Calculate weighted SUVRs for specified regions of interest}
\usage{
calcSUVR(tac, volumes, ROI_def, SUVR_def, corrected = TRUE)
}
\arguments{
\item{tac}{The time-activity curve data from loading function.}

\item{volumes}{The ROI volume data from loading function}

\item{SUVR_def}{is a vector of the start times for window to be used in SUVR}

\item{corrected}{For PVC, true where the data is stored as _C in same tac file.}
}
\value{
A table of SUVR values for the specified ROIs
}
\description{
Calculate weighted SUVRs for specified regions of interest
}
\examples{
calcSUVR(p1tac, p1vol, standardROIs(), c("3000", "3300", "3600", "3900"))
}
