% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSUVR.R
\name{calcSUVR}
\alias{calcSUVR}
\title{Calculate weighted SUVRs for specified regions of interest}
\usage{
calcSUVR(tac, SUVR_def, reference)
}
\arguments{
\item{tac}{The time-activity curve data from calcTAC()}

\item{SUVR_def}{is a vector of the start times for window to be used in SUVR}

\item{reference}{is a string, e.g. "cerbellum", to specify reference region}
}
\value{
A data.frame of SUVR values for the specified ROIs
}
\description{
When smaller ROIs need to be combined into larger ROIs, e.g. when TACs have
been calculated for components of the frontal lobe, but an SUVR is desired
for the entire frontal lobe, the TACs need to be combined, and the relative
volumes of each needs to be taken into account. This function calculates
those weighted means. It also takes into account potentially different
frame durations.
}
\examples{
calcSUVR(p1tac, c(3000, 3300, 3600, 3900))
}
