% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullTAC.R
\name{calcTAC}
\alias{calcTAC}
\title{Calculate weighted time-activity curves for specified regions of interest}
\usage{
calcTAC(tac, volumes, ROI_def, merge = F, PVC = F)
}
\arguments{
\item{tac}{The time-activity curve data from loading function.}

\item{volumes}{The ROI volume data from loading function}

\item{ROI_def}{The definition of ROIs by combining smaller ROIs from TAC file.}

\item{merge}{If true, includes the original ROIs in the output data.}
}
\value{
Time-activity curves for the specified ROIs
}
\description{
Calculate weighted time-activity curves for specified regions of interest
}
\examples{
calcTAC(p1tac, p1vol, standardROIs(), merge=T)
}
