% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_tac}
\alias{load_tac}
\title{Loads TAC from file for use by other functions (default is PMOD .tac format)}
\usage{
load_tac(filename, format = "PMOD", acqtimes = NULL,
  time_unit = NULL, activity_unit = NULL)
}
\arguments{
\item{filename}{(e.g. participant.TAC)}

\item{format}{Options include "PMOD", "voistat" (also from PMOD), and "magia"}

\item{acqtimes}{Filename for a .acqtimes file (as in PMOD), required for 
format="voistat"}

\item{time_unit}{NULL if in file (e.g. PMOD .tac), or set to "seconds" or 
"minutes"}

\item{activity_unit}{NULL if in file (e.g. PMOD .tac), or set to "kBq/cc", 
"Bq/cc", "nCi/cc"}
}
\value{
data.frame with loaded TAC data
}
\description{
Loads TAC from file for use by other functions (default is PMOD .tac format)
}
\examples{
f_raw_tac <- system.file("extdata", "AD06.tac", package="tacmagic") 
tac <- load_tac(f_raw_tac)

}
