% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_tac}
\alias{load_tac}
\title{Loads TAC from file for use by other functions (default is PMOD .tac format)}
\usage{
load_tac(filename, format = "PMOD", acqtimes = NULL)
}
\arguments{
\item{filename}{(e.g. participant.TAC)}

\item{format}{Options include "PMOD", "voistat" (also from PMOD), and "magia"}

\item{acqtimes}{File name for a .acqtimes file (as in PMOD), required for 
format="voistat"}
}
\value{
data.frame with loaded TAC data
}
\description{
Loads TAC from file for use by other functions (default is PMOD .tac format)
}
