% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tac.R
\name{plot_tac}
\alias{plot_tac}
\title{Plots time activity curves from 1 or 2 participants or groups.}
\usage{
plot_tac(TACtable1, TACtable2 = NULL, ROIs, ymax = 25,
  seconds_to_mins = F, title = "")
}
\arguments{
\item{TACtable1}{(e.g. from tac_roi() or load_tac())}

\item{TACtable2}{An optional, second TAC, to plot for comparison}

\item{ROIs}{A vector of ROIs to plot, names matching the TAC headers}

\item{ymax}{The maximum value on the y-axis}

\item{seconds_to_mins}{If true, converts time from TAC from sec to min}

\item{title}{A title for the plot}
}
\value{
Creates a plot.
}
\description{
Plots time activity curves from 1 or 2 participants or groups.
}
