% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tac.R
\name{plotTAC2}
\alias{plotTAC2}
\title{Plots time activity curves from 1 or 2 participants or groups.}
\usage{
plotTAC2(TACtable1, TACtable2 = NULL, ROIs = c("totalcortical",
  "cerebellum"), ymax = 25, seconds_to_mins = FALSE)
}
\arguments{
\item{TACtable1}{(e.g. from calcTAC() or groupTAC(), or simply loadTACfile())}

\item{TACtable2}{An optional, second TAC, to plot for comparison.}

\item{ROIs}{A vector of ROIs to plot, names matching the TAC headers.}

\item{ymax}{The maximum value on the y-axis.}

\item{seconds_to_mins}{If true, converts time from TAC from sec to min.}
}
\value{
Creates a plot.
}
\description{
Plots time activity curves from 1 or 2 participants or groups.
}
