% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROI_definitions.R
\name{roi_ham_stand}
\alias{roi_ham_stand}
\title{Return a list of merged ROIs made up of the atomic ROIs in the Hammer's 
atlas (see references()$Hammers_2003).}
\usage{
roi_ham_stand()
}
\value{
A list of lists, where each list is an ROI (e.g.) frontal lobe that 
specifies the atomic ROIs from the atlas that make it up.
}
\description{
Return a list of merged ROIs made up of the atomic ROIs in the Hammer's 
atlas (see references()$Hammers_2003).
}
\examples{
roi_ham_stand()
}
