% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tacmagic.R
\docType{package}
\name{tacmagic}
\alias{tacmagic}
\alias{tacmagic-package}
\title{tacmagic: PET Time Activity Curve Analysis in R}
\description{
This package provides functions to analyze of PET time activity curve (tac) 
data, including loading from various fomrmats, merging ROIs weighted for 
volume, calculating models including SUVR and DVR, plotting functions and
calculation of cutoff values. Please see the vignette for a detailed 
overview of tacmagic functions.
}
