% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEAR.R
\docType{package}
\name{tacmagic}
\alias{tacmagic}
\alias{tacmagic-package}
\title{tacmagic: PET Time Activity Curve Analysis in R}
\description{
This is a beta package of functions to analyze of PET time activity 
   curve (tac) data, including merging ROIs, calculating SUVR, and plotting. 
   In its current form, these scripts may be useful in analyzing the output 
   files from PMOD or other sources.
}
\section{Foo functions}{

The foo functions ...
}

